Split `MessageView` into `ChatView` and `MessageView` `ChatView` works with a full chat. `MessageView` works on a single `Message`.
